- Use u_int instead of unsigned long for storing   vm related values returned by sysctl - Minor style fix in preproccessor comments
git-svn-id: https://conky.svn.sourceforge.net/svnroot/conky/trunk/conky1@1215 7f574dfc-610e-0410-a909-a81674777703

+++ src/freebsd.c
@@ -146,7 +146,7 @@ int check_mount(char *s)
 
 void update_meminfo()
 {
-	unsigned long total_pages, inactive_pages, free_pages;
+	u_int total_pages, inactive_pages, free_pages;
 	unsigned long swap_avail, swap_free;
 
 	int pagesize = getpagesize();
